fixup! [HACK, WIP]: Insert runc into tether

+++ lib/tether/tether.go
@@ -808,41 +808,44 @@ func (t *tether) launch(session *SessionConfig) error {
 	log.Debugf("Resolved %s to %s", session.Cmd.Path, resolved)
 	session.Cmd.Path = resolved
 
-	// [HACK]: Defer cleanup
-	if _, err := os.Stat("config.json"); err == nil {
-		panic("hacky code should not be overwriting files")
+	// [HACK]: This is a terrible condition; better to express this as a configurable attribute of a session(?)
+	if _, err := os.Stat("/usr/sbin/runc"); err == nil {
+		// [HACK]: Defer cleanup
+		if _, err := os.Stat("config.json"); err == nil {
+			panic("hacky code should not be overwriting files")
+		}
+
+		defer os.Remove("config.json")
+
+		// [HACK]: Write runc config
+		specCmd := exec.Command("/usr/sbin/runc", "spec")
+		err = specCmd.Start()
+		if err != nil {
+			panic("error generating spec") // TODO[HACK]: handle errors
+		}
+
+		// [HACK]: Tell runc to run `resolved`
+		setCmd := exec.Command("/bin/bash", "-c", "cat config.json | jq '.process.args = [\""+strings.Join(session.Cmd.Args, "\",\"")+"\"]' > config.json.updated") // TODO[HACK]: 😂
+		err = setCmd.Start()
+		if err != nil {
+			panic("error setting binary") // TODO[HACK]: handle errors
+		}
+
+		os.Rename("config.json.updated", "config.json") // [HACK]: Atomic update
+
+		// [HACK]: Tell runc the rootfs is in the parent directory
+		setRootFS := exec.Command("/bin/bash", "-c", "cat config.json | jq '.root.path = \"../\"' > config.json.updated") // TODO[HACK]: 😂
+		err = setRootFS.Start()
+		if err != nil {
+			panic("error setting rootfs") // TODO[HACK]: handle errors
+		}
+
+		os.Rename("config.json.updated", "config.json") // [HACK]: Atomic update
+
+		// [HACK]: Replace `session.Cmd.Path` with `runc run`
+		session.Cmd.Path = "/usr/sbin/runc"
+		session.Cmd.Args = []string{"run", "-b", ".", session.Name}
 	}
-
-	defer os.Remove("config.json")
-
-	// [HACK]: Write runc config
-	specCmd := exec.Command("/usr/sbin/runc", "spec")
-	err = specCmd.Start()
-	if err != nil {
-		panic("error generating spec") // TODO[HACK]: handle errors
-	}
-
-	// [HACK]: Tell runc to run `resolved`
-	setCmd := exec.Command("/bin/bash", "-c", "cat config.json | jq '.process.args = [\"" + strings.Join(session.Cmd.Args, "\",\"") + "\"]' > config.json.updated") // TODO[HACK]: 😂
-	err = setCmd.Start()
-	if err != nil {
-		panic("error setting binary") // TODO[HACK]: handle errors
-	}
-
-	os.Rename("config.json.updated", "config.json") // [HACK]: Atomic update
-
-	// [HACK]: Tell runc the rootfs is in the parent directory
-	setRootFS := exec.Command("/bin/bash", "-c", "cat config.json | jq '.root.path = \"../\"' > config.json.updated") // TODO[HACK]: 😂
-	err = setRootFS.Start()
-	if err != nil {
-		panic("error setting rootfs") // TODO[HACK]: handle errors
-	}
-
-	os.Rename("config.json.updated", "config.json") // [HACK]: Atomic update
-
-	// [HACK]: Replace `session.Cmd.Path` with `runc run`
-	session.Cmd.Path = "/usr/sbin/runc"
-	session.Cmd.Args = []string{"run", "-b", ".", session.Name}
 
 	// block until we have a connection
 	if session.RunBlock && session.ClearToLaunch != nil {
